[feature] update kb api, support multiple kb_name and collection_name

+++ libs/rag/server/kb/milvus_kb_service.py
@@ -25,7 +25,7 @@ class MilvusKBService(KBService):
         embedding_model_config: ModelConfig = get_model_configs(embed_model)
         embed_dim = embedding_model_config.meta_data.get("embed_size", 1024)
         self.embed_func = PlatformLLM(embedding_model_config).embed
-        if kb_info is None:
+        if kb_info is None or len(kb_info.strip()) == 0:
             kb_info = f"Milvus KB Service, based on {embed_model}, dim {embed_dim}"
         super().__init__(kb_name, kb_info, embed_model)
         self._DEFAULT_FIELDS = [
@@ -80,9 +80,13 @@ class MilvusKBService(KBService):
             data=[query_embedding],
             search_params=search_params,
             limit=top_k,
-            output_fields=["metadata", "content"],
+            output_fields=["id", "distance", "metadata", "content"],
         )[0]
-        contexts = [Context.model_validate(obj["entity"]) for obj in results]
+        results = [r for r in results if r["distance"] >= score_threshold]
+        results = [
+            {"id": r["id"], "distance": r["distance"], **r["entity"]} for r in results
+        ]
+        contexts = [Context.model_validate(obj) for obj in results]
         return contexts
 
     def create_collection(
@@ -121,7 +125,8 @@ class MilvusKBService(KBService):
         self.client.release_collection(collection_name)
 
     def drop_collection(self, collection_name):
-        self.client.drop_collection(collection_name)
+        if collection_name in self.list_collection():
+            self.client.drop_collection(collection_name)
 
     def add_context(self, collection_name: str, context: Union[Context, List[Context]]):
         if isinstance(context, Context):
@@ -136,7 +141,7 @@ class MilvusKBService(KBService):
     def get_obj_by_uuid(self, collection_name: str, uuid: str):
         return self.client(collection_name=collection_name, ids=[uuid])
 
-    def do_create_kb(self, kb_name, kb_info, embed_model):
+    def do_create_kb(self, kb_name: str, kb_info: str, embed_model: str):
         pass
 
     def save_vector_store(self):
